+++ packages/smoldot/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+### Added
+
+- Added `@polkadot-api/smoldot/node-worker` to create a smoldot worker in NodeJS.
+- Added `@polkadot-api/smoldot/from-node-worker` to create a client from a node smoldot worker.
+
 ## 0.2.7 - 2024-07-25
 
 ### Fixed